Convert EXPR_ and TYP_ to polymorphic variants.
Preparation for removing the disgusting feature of the
front end in which type expression are first value expressions
and then have to be translated. This means introducing constructors
like EXPR_rptsum_type which are translated to TYP_reptsum_type
but we're left with a constructor of no further use that has
to be dynamically "assert false"d out in pattern matches
to retain exhaustion.

Using polymorphic variants I hope to merge the TYP_ and EXPR_
constructors and eliminate the conversion. Then the dynamic
assert falsing can still be done, but can be done on a whole
group of constructors at once. Extension will then be easier
to manage (requiring less edits to ensure pattern matching
exhaustion). At least .. that's the idea :)