Reduce pointer representation during monomorphisation.
At least, I tried to do this. Not sure if I did it right
because at present there are no test cases for which
the old model will fail. So I need to try to prove
it works.

Unification remains problematic. I still don't understand
how to unify the pointer representations correctly.