Replace sync scheduler list of fthreads with an object containing them.
The object also contains a thread count, in preparation
for handling sharing by enabling locking to protect
critical sections if and only if more than one pthread
is sharing the list.

The problem yet to be solved is how to enusre all the
threads remain active until there's nothing left to do.