Simplify and massively speed up schannel read/write.
Previously, an schannel consisted of two slists, one of
which was always empty. An slist is a traditional linked
list of nodes, with the data being a pointer to an fthread.

The new schannel is a single pointer with the low bit
set if it points to a writer,  nullptr if empty, and
low bit clear but not null if it pointers to a reader.
The list of fthreads is now maintened by an auxilliary
pointer in the fthread object.

Consequently schannels are now one machine word,
and no allocations (and hence no deallocation later by
the GC) is required to manage them, other than
the usual allocation and garbage collection of the
fthread objects themselves.

So it should be a LOT faster! And also eliminates the
use of slists.