Add linear function types.
Also effectors, and procedures as well.
Carry linearity through to code generator, in case C++ might
happen to be able to use linearity optimisations. Maybe get rid
of this later.

A cast is needed on a closure at the moment, there's not yet
syntax to define a linear function.

linear function types are subtypes of ordinary function types.
A function accepting a nonlinear function type should work
if given a linear function.

linear functions are ordinary functions, they just happen
to promise to treat their argument "as if" it were unique.

The linearity of a function could be checked by the control
flow analysing uniqueness checker by a trick: if the function
is linear, treat it as a function with a unique parameter type.

However note that *calls* to the function treat the domain
as non-unique.

The problem with this is that late uniqueness checking may
be defeated by inlining. In other words, if a linear function
is inlined into a non-linear one, we've lost the information
required to do the trick. However an early check should
work, and we do currently do TWO checks, one early and
one late, in particular we do a check before monomorphisation.