Make Felix recognise ObjC NSString @"literals".
Note there is not as yet any library support.
The user must somehow require package "foundation" for it to work.
All Felix literal styles are supported, single and double
quote marks and the same again tripled.
There's no raw option. Felix escape encoding is performed as usual.
The output in C++/ObjC++ will be UTF8 encoded.
Felix does not support UCS2 or UCS4 encoding.
This could be wrong for ObjC, not sure.

The flxg compiler does NOT perform any const folding
on this type at this time.