Add $i code which is replaced by contents of sole string argument.
It's a hack which may be fixed a bit soon.
We can now define this:

fun selector : string -> SEL = '@@selector($i)';

and then this code:

  selector "Hello"

is replaced by

  @selector(Hello)

Actually we want the input in Felix to be

  @selector(Hello)

too. The implementation only works if the input is a string
literal AND whilst the library is mapping string literals
to ::std::string("literal here"), which is real hack.
The parser can, in theory, implement and enforce the meaning
of @selector(identifier). the test code uses:

fun selector : string -> SEL = '@@selector($i)';

and the $i feature in the compiler's csubst.ml routine
strips the ::std::string(" off the front and ") off the back.
Hacky. There's a better way (identifier literals) but it requires
more pervasive changes to the compiler.