Change the semantics for ivar projections.
We use a phantom proxy type to allow an opaque
type intended to emulate a pointer to admit
value and pointer projections in our bindings
for ivars in Objective C. In particular for an
Objective C type like NSString, we provide two
bindings, one for Felix type NSString to C type
NSString*, and one for Felix type _derefproxy_NSString
to void. This type is never used in the generated code
but a real type is required because due to an issue
a typedef is generated.

We define the deref operator on the Felix type to return
a proxy. Then we can overload so an ivar name acts as a
pointer projection applied to the actual type, and a value
projection applied to the deref proxy. This means

  v . name

is a machine pointer to the address of the ivar name
in the instance v, whilst

  v -> name

is mapped by the parser to the same as v *. name which is

  (*v) . name

and since *v is now a proxy type the projection overload
of name returns the value of ivar name in the instance v.
If the client writes this:

  var bad = *v;

it will unfortunately PASS Felix type checking but since
the proxy is type void the generated C++ will be

  void bad = whatever;

which will fail C++ type checks. This could be fixed with
a special property so the error is caught be Felix at
code generation time I think.

The machinery can also be used with an apply() function
to make varray work with sane syntax:

  v . 42 // pointer projection
  v -> 42 // value projection