Add new symbols for callback types.
If name is the name of the callback thunk, then:

1. name'c_callback_type is the type C expects with void *client_data
2. name'felix_callback_type is the type of the Felix function
which is passed to the thunk, the client data argument is removed
3. name'felix_thunk_type is the type of the thunk

To pass the thunk to C, the thunk has to be cast to
the c_callback_type

To pass the Felix callback to C as client data it has to be
cast type address (void* in C)

The typing is cast free and properly type checked
if the callback thunk and Felix callback
are passed to a Felix event loop, i.e. the thunk expects
a value of the felix callback type.

The decision to favour correctness over the most common
case is dubious. Almost all callbacks will be used
interfacing with C code. In Felix, the callback can be
called directly without needing a thunk.

Therefore this design might be changed, it's just a bit
of a pain to make the thunk code cast a void * to the
felix callback type internally before invoking it.
At present the thunk code doesn't need the cast because
the parameter type is already correct.

At least now, the relevant types can be found easily
without having to manually write them out as well
as defining the callback thunk.