Rework RTL.
Fibres now contain a pointer to their owning fthread_list.
Spinlocks reworked and applied to read/write operations.
Schannels should now work across pthreads correctly.
Fibres are permanently associated with a particular fthread list,
and so can no longer migrate.

Conditional spinlocking is removed since it is no longer
possible to tell when a second thread is launched with
spawn_fibre or spawn_process since the thread control
object is not reachable from fibres. Spinlocks are always used
now, which is simpler and safer, and may not be much slower for
the single threaded case, and will be faster for the many threads
since the test of the condition is removed.

Implementing fibre affinity for a fixed active list has much
simpler more comprehensible semantics.