Make Felix build with latest Ocaml.
Dypgen has been hacked to not emit line numbers.
This is unfortunate but the code used String.t mutations
which are no longer supported. The code could have been
properly fixed to still support line numbers, but removing
the feature just seemed easier than trying to constantly switch
between string and bytes types. Some of that had to be done anyhow.

As a result, any errors introduced into the bootstrap parser
of Felix, flx_parse.mly, will now show a fault in the generated
flx_parse.ml file instead of the mly file as it used to.

This file is not changed very often and it would now be harder
to debug a mistake in a change but not impossible.