Root fthreads constructed by circuit statement.
This is essential in RT threads so that the fthreads and channels
created by a circuit are reachable, so that a garbage collection
for which the RT thread does not stop does not delete these components,
even if other paths to them are changing during collection.
RT threads do not stop for a GC.

Note: we postulate that the mark phase of the GC will still
terminate despite the changing paths. However RT threads should not
allocate any new GC objects, as the thread stack is never scanned.

The procedure starting the main fibre of an RT thread must
also always be reachable. It's not clear how to do this
at the moment because it is possible it is cloned.
Normally procedures are linked together by a spaghetti stack,
however the initial procedure of a thread is converted
to a continuation which is vested in a fibre run by the
threads main scheduler: we need to be sure that scheduler
is reachable too.