Remove -fblocks from posix build flags in bootstrap build.
Somehow this gets applied to g++ which doesn't support it.
The flag only applies to clang I think. I don't know if
it is available on Linux when using clang.

I've left the flag out of the darwin spec, so the bootstrap
is back to its old state. We'll see if anyone gets a compiler
error without the flag. The flag is set for clang_macos in
the real (felix generated) toolchains.