Add almost working type function code.
The binding is reworked to strict phases.
Beta reduction is done on everything (supposedly) at the end.
All the tests pass except the GUI which all fail with what
appears to be an unreduced term. Unfortunately, the failure
is at an extremely low level and there are no tracking diagnostics.

So I will commit what I have since most of it works,
and it self-rebuilds ok. I have missed beta-reducing something.