Rewrite alpha and beta. Remove expand pass next because it doesn't work.
The problem with the expansion is that a recursive call is unrolled once
because the expansionn just expands a term starting with an empty trail,
so only the second call (in the first unfolding) because a fixpoint.

Previously that failed by actually setting a fixpoint where the recursion
was polymorphic. This is because I forgot to alpha convert the function
before substitution. If I'd done that, the polymorphic recursion
would have expanded forever.

In Fact, the expansion is unbounded, but it occurs inside a typematch
on only one branch so would be terminated by the other branch being
taken. Exactly like an ordinary function.