Beta reduce argument of application before substitution.
Probably this is wrong. If the argument is an application
returning a type tuple required by the function, the correct
method would be to apply projections to the unreduced
argument however we have no such projection terms at this time,
although the could be defined with a typematch.

ALL operations (eager reduction, projections, lazy evaluation)
will lead to serious issues with fixpoints. But we have no
failing test cases at the moment.