Add package channel.hpp which contains most of the old code from the nurtl repository. Everything added compiles.
However there is a known BUG: fibres MUST have a registry of channels.
When I/O is done, ALL the channels in the registry have to have
their reference count decremented. On resume, incremented.

The current code only decrements the refcount of the cnannel
on which the request is made. This is not enough.